Fix 404 when manually changing URL in browser and after redirects
The users are redirected from Auth0 to the callback URL and we get an error saying the page can’t be found.
This is because BrowserSync thinks we’re requesting the callback/ directory, and it can’t find the index.html
file in that directory - because it doesn’t exist. To fix this  behaviour we should add
the connect-history-api-fallback middleware